esr: unify replay numeric types to double and clean conversion warnings

+++ include/1q/electronic_surveillance_radar/model/EmitterObservation.h
@@ -29,12 +29,12 @@ enum class EsrObservationQuality {
 struct ONEQ_API EmitterObservation {
   std::uint64_t observation_id{0U};                     /**< 观测记录唯一标识 */
   double timestamp_s{0.0};                              /**< 观测时间戳（单位：s） */
-  float aoa_az_deg{0.0f};                               /**< 测得方位角（单位：deg） */
-  float aoa_el_deg{0.0f};                               /**< 测得俯仰角（单位：deg） */
+  double aoa_az_deg{0.0};                               /**< 测得方位角（单位：deg） */
+  double aoa_el_deg{0.0};                               /**< 测得俯仰角（单位：deg） */
   double rf_hz{0.0};                                    /**< 测得载频（单位：Hz） */
   double pulse_width_s{0.0};                            /**< 测得脉宽（单位：s） */
-  float amplitude_db{0.0f};                             /**< 接收幅度（单位：dB） */
-  float snr_db{0.0f};                                   /**< 观测信噪比（单位：dB） */
+  double amplitude_db{0.0};                             /**< 接收幅度（单位：dB） */
+  double snr_db{0.0};                                   /**< 观测信噪比（单位：dB） */
   EsrObservationQuality quality{EsrObservationQuality::kLow}; /**< 观测质量等级 */
   bool is_jammed{false};                                /**< 该观测是否受干扰显著影响 */
 };
